+++ src/clack/__init__.py
@@ -8,6 +8,7 @@ import logging as _logging
 
 from . import xdg
 from ._config import Config
+from ._config_file import ConfigFile, YAMLConfigFile
 from ._helpers import NewCommand, comma_list_or_file, new_command_factory
 from ._main import MainType, main_factory
 from ._parser import Parser, filter_cli_args
@@ -15,9 +16,11 @@ from ._parser import Parser, filter_cli_args
 
 __all__ = [
     "Config",
+    "ConfigFile",
     "MainType",
     "NewCommand",
     "Parser",
+    "YAMLConfigFile",
     "comma_list_or_file",
     "filter_cli_args",
     "main_factory",