+++ tests/test_clack.py
@@ -1,9 +1,11 @@
 """Miscellaneous tests for the clack library."""
 
+from eris import Err
 import pytest
 
 import clack
 from clack import _dynvars as dyn
+from clack.pytest_plugin import MakeConfigFile
 
 from .shared import Config
 
@@ -31,3 +33,26 @@ def test_config_is_immutable() -> None:
 
         with pytest.raises(TypeError):
             cfg.do_stuff = False
+
+
+def test_config_file(make_config_file: MakeConfigFile) -> None:
+    """Test the clack.ConfigFile protocol implementations."""
+    cf = make_config_file("clack.yml", foo="FOO", bar=3, baz=False)
+
+    assert sorted(cf.to_dict().unwrap().items()) == [
+        ("bar", 3),
+        ("baz", False),
+        ("foo", "FOO"),
+    ]
+
+    assert cf.get("foo").unwrap() == "FOO"
+    assert cf.set("foo", "KUNG").unwrap() == "FOO"
+    assert isinstance(cf.set("fool", "FOOL"), Err)
+
+    assert cf.set("fool", "FOOL", allow_new=True).unwrap() is None
+    assert sorted(cf.to_dict().unwrap().items()) == [
+        ("bar", 3),
+        ("baz", False),
+        ("foo", "KUNG"),
+        ("fool", "FOOL"),
+    ]